+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 Yii2-ActiveRecord-History Extension for Yii 2 ChangeLog
 ==============================================
 
+1.2.1 Jul 05, 2022
+-------------------------
+
+- Added method to retrieve the value of a model attribute at a given date-time
+
 1.2 Jun 13, 2019
 -------------------------
 
-- Optain the last change of a model attribute
-- Optain all the changes of a model
+- Obtain the last change of a model attribute
+- Obtain all the changes of a model
 - Add docs
 
 1.1.2.1 Oct 13, 2015